Move conversation sandbox owner map to adapter
